backup_old/requirements.txt

Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Dance==7.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
Werkzeug==2.3.6
psycopg2-binary==2.9.7
SQLAlchemy==2.0.19
Pillow==10.0.0
beautifulsoup4==4.12.2
mss==9.0.1
cryptography==41.0.3
paypalrestsdk==1.13.1
opencv-python==4.8.1.78
# Database and ORM
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
# Authentication and Authorization
flask-login==0.6.3
flask-dance==7.0.0
# Security
flask-wtf==1.2.1
flask-talisman==1.1.0
# Environment and configuration
python-dotenv==1.0.1
openai>=0.27.0
aiohttp
numpy
psutil
pyodbc
yt-dlp
aiohttp
beautifulsoup4
cryptography
Flask
flask-dance
flask-login
flask-migrate
flask-sqlalchemy
flask-talisman
flask-wtf
google-auth
google-auth-httplib2
google-auth-oauthlib
mss
numpy
openai
opencv-python
paypalrestsdk
Pillow
psutil
psycopg2-binary
pyodbc
python-dotenv
requests
requests-oauthlib
SQLAlchemy
Werkzeug
yt-dlp
